requirements.txt
asgiref==3.5.2
astroid==2.12.13
black==22.12.0
blurhash==1.1.4
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rdkit @ git+https://github.com/rdkit/django-rdkit.git@cd1d69af3fb024dae1c98c939cde07f6a0583751
django-rest-knox==4.2.0
django-reversion==5.0.4
django-stubs==1.13.1
django-stubs-ext==0.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-stubs==1.8.0
drf-access-policy==1.3.0
filelock==3.8.2
gunicorn==20.1.0
identify==2.5.11
idna==3.3
isort==5.11.4
lazy-object-proxy==1.8.0
maturin==0.13.2
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.2
oauthlib==3.2.0
pathspec==0.10.3
Pillow==9.2.0
platformdirs==2.6.0
pre-commit==2.21.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
rdkit==2022.3.5
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.6
types-pytz==2022.7.0.0
types-PyYAML==6.0.12.2
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.17.1
wrapt==1.14.1
